refactor : 전체적인 코드를 리펙토링하였습니다.
- 기존 레거시 코드를 객체지향원칙을 준수하며 리펙토링
- 책임이 한 곳에 몰려있는 것을 분산 시킴
- 각 기능별 패키지 구조 개선
