--- src/main/java/com/project/musinsastocknotificationbot/crawling/Crawling.java
@@ -1,13 +1,0 @@
-package com.project.musinsastocknotificationbot.crawling;
-
-import com.project.musinsastocknotificationbot.crawling.dto.TrackProductRequest;
-import com.project.musinsastocknotificationbot.crawling.dto.TrackProductResponse;
-import com.project.musinsastocknotificationbot.product.dto.Response.CrawlingResponse;
-import java.util.List;
-
-public interface Crawling {
-
-  public CrawlingResponse crawling (String url);
-
-  List<TrackProductResponse> track(List<TrackProductRequest> trackProductRequests);
-}
