+++ src/main/java/com/project/musinsastocknotificationbot/domain/product/error/JsoupIOException.java
@@ -1,0 +1,7 @@
+package com.project.musinsastocknotificationbot.domain.product.error;
+
+public class JsoupIOException extends RuntimeException {
+
+  public JsoupIOException(Exception e) {
+  }
+}
