+++ src/main/java/com/project/musinsastocknotificationbot/domain/product/event/ProductEvent.java
@@ -1,0 +1,11 @@
+package com.project.musinsastocknotificationbot.domain.product.event;
+
+public record ProductEvent(
+    String message
+) {
+
+  public static ProductEvent of(String message) {
+    return new ProductEvent(message);
+  }
+}
+
