+++ src/main/java/com/project/musinsastocknotificationbot/domain/product/repository/ProductRepository.java
@@ -1,0 +1,8 @@
+package com.project.musinsastocknotificationbot.domain.product.repository;
+
+import com.project.musinsastocknotificationbot.domain.product.entity.Product;
+import com.project.musinsastocknotificationbot.domain.product.entity.vo.ProductInfo;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface ProductRepository extends JpaRepository<Product, ProductInfo> {
+}
