+++ src/main/java/com/project/musinsastocknotificationbot/infrastructure/crawling/dto/TrackProductRequest.java
@@ -1,0 +1,13 @@
+package com.project.musinsastocknotificationbot.infrastructure.crawling.dto;
+
+import com.project.musinsastocknotificationbot.domain.product.entity.Product;
+
+public record TrackProductRequest (
+    Long id,
+    String size
+) {
+
+  public TrackProductRequest(Product product) {
+    this(product.getProductId().getId(), product.getProductId().getSize());
+  }
+}
