+++ src/main/java/com/project/musinsastocknotificationbot/infrastructure/message/service/MessageService.java
@@ -1,0 +1,6 @@
+package com.project.musinsastocknotificationbot.infrastructure.message.service;
+
+public interface MessageService {
+
+  public void sendMessage(String message);
+}
