--- src/main/java/com/project/musinsastocknotificationbot/message/service/MessageService.java
@@ -1,6 +1,0 @@
-package com.project.musinsastocknotificationbot.message.service;
-
-public interface MessageService {
-
-  public void sendMessage(String message);
-}
