--- src/main/java/com/project/musinsastocknotificationbot/product/domain/repository/ProductRepository.java
@@ -1,8 +1,0 @@
-package com.project.musinsastocknotificationbot.product.domain.repository;
-
-import com.project.musinsastocknotificationbot.product.domain.Product;
-import com.project.musinsastocknotificationbot.product.domain.vo.ProductInfo;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface ProductRepository extends JpaRepository<Product, ProductInfo> {
-}
