--- src/main/java/com/project/musinsastocknotificationbot/product/error/JsoupIOException.java
@@ -1,7 +1,0 @@
-package com.project.musinsastocknotificationbot.product.error;
-
-public class JsoupIOException extends RuntimeException {
-
-  public JsoupIOException(Exception e) {
-  }
-}
