--- src/main/java/com/project/musinsastocknotificationbot/product/event/ProductEvent.java
@@ -1,11 +1,0 @@
-package com.project.musinsastocknotificationbot.product.event;
-
-public record ProductEvent(
-    String message
-) {
-
-  public static ProductEvent of(String message) {
-    return new ProductEvent(message);
-  }
-}
-
