--- src/main/java/com/project/musinsastocknotificationbot/product/service/ProductService.java
@@ -1,12 +1,0 @@
-package com.project.musinsastocknotificationbot.product.service;
-
-import com.project.musinsastocknotificationbot.product.domain.vo.ProductInfo;
-
-public interface ProductService {
-
-    void save(ProductInfo productInfo);
-
-    void findAll();
-
-    void delete(ProductInfo productInfo);
-}
